Revert "Lesson2 sensors"
