+++ app/src/main/java/ru/vasic2000/newweather/Fragments/Game_1.java
@@ -1,0 +1,39 @@
+package ru.vasic2000.newweather.Fragments;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import ru.vasic2000.newweather.Activities.MainActivity;
+import ru.vasic2000.newweather.R;
+
+public class Game_1 extends Fragment {
+    private TextView tv_goBack1;
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_game_1, container, false);
+    }
+
+    @Override
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        View game1 = getView();
+
+        tv_goBack1 = game1.findViewById(R.id.tv_return1);
+
+        tv_goBack1.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                MainActivity ma = (MainActivity) getActivity();
+                ma.fragmentBack();
+            }
+        });
+    }
+}
