+++ app/src/main/java/ru/vasic2000/newweather/Fragments/Game_3.java
@@ -1,0 +1,39 @@
+package ru.vasic2000.newweather.Fragments;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import ru.vasic2000.newweather.Activities.MainActivity;
+import ru.vasic2000.newweather.R;
+
+public class Game_3 extends Fragment {
+    private TextView tv_goBack3;
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_game_3, container, false);
+    }
+
+    @Override
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        View game3 = getView();
+
+        tv_goBack3 = game3.findViewById(R.id.tv_return3);
+
+        tv_goBack3.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                MainActivity ma = (MainActivity) getActivity();
+                ma.fragmentBack();
+            }
+        });
+    }
+}
