+++ app/src/main/res/values-xhdpi/dimens.xml
@@ -11,7 +11,4 @@
     <dimen name="MargineTopFirst">10dp</dimen>
     <dimen name="MargineLeftFirst">10dp</dimen>
     <dimen name="indicatorSize">400dp</dimen>
-    <dimen name="Humid_Size">25sp</dimen>
-    <dimen name="humid_value_font_size">40sp</dimen>
-    <dimen name="temper_value_font_size">40sp</dimen>
 </resources>
