+++ app/src/main/res/values/strings.xml
@@ -44,14 +44,5 @@
     <string name="telegramm">telegramm: @Vasyator</string>
     <string name="about_developer">About developer</string>
     <string name="feedback2">Feedback</string>
-    <string name="sensors">Sensors</string>
-    <string name="title_temperature">Temperature sensor</string>
-    <string name="title_humidity">Relative humidity sensor</string>
-    <string name="sensors2">Sensors</string>
-    <string name="temperature2">Temperature</string>
-    <string name="humidity2">Humidity</string>
-    <string name="relative_humidity">Relative humidity</string>
-    <string name="temperature">Temperature</string>
-    <string name="no_sensor">No sensor!</string>
 
 </resources>
